1.11.8.7

+++ src/main/java/fr/xyness/SCS/API/SimpleClaimSystemAPI_Impl.java
@@ -218,7 +218,7 @@ public class SimpleClaimSystemAPI_Impl implements SimpleClaimSystemAPI {
 	}
 
 	@Override
-	public boolean addClaimSale(Claim claim, double claimPrice) {
+	public boolean addClaimSale(Claim claim, long claimPrice) {
 		return instance.getMain().setChunkSale(claim, claimPrice).join();
 	}
